fix: resolved error creating a routine
