+++ lib/app/modules/routine/bindings/routine_binding.dart
@@ -1,0 +1,12 @@
+import 'package:get/get.dart';
+
+import '../controllers/routine_controller.dart';
+
+class RoutineBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<RoutineController>(
+      () => RoutineController(),
+    );
+  }
+}
