+++ lib/app/routes/app_pages.dart
@@ -1,8 +1,10 @@
 import 'package:get/get.dart';
-import 'package:routine/app/modules/home/views/routine_create_view.dart';
+import 'package:routine/app/modules/routine/views/routine_create_view.dart';
 
 import '../modules/home/bindings/home_binding.dart';
 import '../modules/home/views/home_view.dart';
+import '../modules/routine/bindings/routine_binding.dart';
+import '../modules/routine/views/routine_view.dart';
 
 part 'app_routes.dart';
 
@@ -20,7 +22,12 @@ class AppPages {
     GetPage(
       name: _Paths.CREATE,
       page: () => const RoutineCreateView(),
-      binding: HomeBinding(),
+      binding: RoutineBinding(),
+    ),
+    GetPage(
+      name: _Paths.ROUTINE,
+      page: () => const RoutineView(),
+      binding: RoutineBinding(),
     ),
   ];
 }
