+++ lib/app/routes/app_routes.dart
@@ -5,10 +5,12 @@ abstract class Routes {
   Routes._();
   static const HOME = _Paths.HOME;
   static const CREATE = _Paths.CREATE;
+  static const ROUTINE = _Paths.ROUTINE;
 }
 
 abstract class _Paths {
   _Paths._();
   static const HOME = '/home';
   static const CREATE = '/create';
+  static const ROUTINE = '/routine';
 }
